Разбил логику кода на отдельные части (классы) с разной ответственностью. Связал классы между собой через композицию. Приложение запускается, все работает, но мне не нравится, что рандомные вопросы повторяются периодически
