+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,9 @@
+
+import Foundation
+
+//Вью модель для вопросов
+            struct QuizQuestion {
+                let image: String
+                let text: String
+                let correctAnswer: Bool
+            }
