+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,9 @@
+
+import Foundation
+
+//Вью модель для состяния "Результат квиза"
+                struct QuizResultsViewModel {
+                    let title: String
+                    let text: String
+                    let buttonText: String
+                }
