+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,9 @@
+
+import UIKit
+
+//Вью модель для состояния "Вопрос задан"
+                struct QuizStepViewModel {
+                    let image: UIImage
+                    let question: String
+                    let questionNumber: String
+                }
