+++ MovieQuiz/Services/QuestionFactoryProtocol.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+protocol questionFactoryProtocol {
+    func requestNextQuestion() -> QuizQuestion?
+}
